feat(ti): add message to encrypt tfa during suspend
At suspend, BL31 with its context will be encrypted by TIFS in DDR.
Encryption is needed for security matters, so that the BL31 is not
modified before entering suspend or early at resume.

We only need the encryption function here because the decryption message
will be send by the R5 SPL at resume.

Also introduce the LPM_ENCRYPT_IMAGE cap signals that FW has the support
to encrypt the image using the TISCI_MSG_LPM_ENCRYPT tisci message.

This is useful in suspend to ram cases where we would like to
store the encrypted image of a secure fw instead of the original image
itself in the DDR.

Check for LPM_ENCRYPT_IMAGE flag in the FW capabilities, and only then
call encrypt.

Upstream-Status: Pending

Change-Id: I266472da87dd0821493019b2d9853f8886f33811
Signed-off-by: Richard Genoud <richard.genoud@bootlin.com>
Signed-off-by: Abhash Kumar Jha <a-kumar2@ti.com>