feat(k3): choose cluster_start_id depending on the soc
The CLUSTER_DEVICE_START_ID denotes the device id of the A-core cluster.
It is utilized when powering off the entire cluster.

J7200, J721E and J721S2 have a different cluster_start_id than their
"generic" counterparts.

Query the JTAG_ID register to get the part id and choose the
cluster_start_id depending on that.

Upstream-Status: Pending

Change-Id: I44d3ac0ec646c39019e4c0167d34f410015a147a
Signed-off-by: Abhash Kumar Jha <a-kumar2@ti.com>